+++ app/OpenAPI/Schemas/MenuExplore.php
@@ -18,7 +18,7 @@ class MenuExplore
     #[OAT\Property(type: 'array', items: new OAT\Items(type: 'object', properties: [
         new OAT\Property(type: 'string', property: 'name', example: 'Category name'),
         new OAT\Property(type: 'array', property: 'cocktails', items: new OAT\Items(type: 'object', properties: [
-            new OAT\Property(type: 'string', property: 'sort', example: 1),
+            new OAT\Property(type: 'integer', property: 'sort', example: 1),
             new OAT\Property(type: 'object', property: 'price', properties: [
                 new OAT\Property(type: 'integer', property: 'full', example: 2385),
                 new OAT\Property(type: 'string', property: 'formatted', example: 'EUR 23.85'),